+++ src/test/java/com/ryan/roomreservationservice/service/member/MemberServiceImpTest.java
@@ -9,13 +9,11 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.test.context.ActiveProfiles;
 
 import java.util.Optional;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.*;
 
 @ActiveProfiles("memory")
@@ -65,9 +63,10 @@ class MemberServiceImpTest {
 
         Optional<Member> emptyMember = Optional.empty();
 
-        String encryptionPassword = this.passwordEncoder.encode(signUpRequest.getPassword());
+        String encryptionPassword = "encrypted" + signUpRequest.getPassword();
 
         when(this.memberQueryImp.findByEmail(signUpRequest.getEmail())).thenReturn(emptyMember);
+        when(this.passwordEncoder.encode(signUpRequest.getPassword())).thenReturn(encryptionPassword);
         when(this.memberStoreImp.save(signUpRequest.getEmail(), encryptionPassword, signUpRequest.getName(), signUpRequest.getPhone())).thenReturn(member);
 
         // when(실행): 어떠한 함수를 실행하면
@@ -81,17 +80,4 @@ class MemberServiceImpTest {
                         signUpRequest.getPhone());
     }
 
-    @Test
-    public void 비밀번호_암호화_검증() {
-        // given(준비): 어떠한 데이터가 준비되었을 때
-        PasswordEncoder passwordEncoder = new BCryptPasswordEncoder();
-        String password = "1234qweR!!";
-
-        // when(실행): 어떠한 함수를 실행하면
-        String encryptionPassword = passwordEncoder.encode(password);
-        boolean matchesPassword = passwordEncoder.matches(password, encryptionPassword);
-
-        // then(검증): 어떠한 결과가 나와야 한다.
-        assertThat(matchesPassword).isTrue();
-    }
 }
